requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-rest-authtoken==2.1.4
django-versatileimagefield==2.2
djangorestframework==3.13.1
phonenumbers==8.12.51
Pillow==9.1.1
python-magic==0.4.27
pytz==2022.1
sqlparse==0.4.2
